revert config_entry move
